% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossgrid.R
\name{lossgrid}
\alias{lossgrid}
\alias{lossgrid.default}
\alias{lossgrid.penAFT.cva}
\title{lossgrid.penAFT.cva
Cross-validated loss for each value of alpha and lambda parameters from Elastic Net fit}
\usage{
lossgrid(x, ...)

\method{lossgrid}{default}(x, ...)

\method{lossgrid}{penAFT.cva}(x, ..., cv.type = c("min", "1se"))
}
\arguments{
\item{x}{Object of}

\item{cv.type}{This can be either \code{"min"} which is the minimum loss, or \code{"1se"} which is the highest loss within 1 standard error of the minimum.}
}
\description{
lossgrid.penAFT.cva
Cross-validated loss for each value of alpha and lambda parameters from Elastic Net fit
}
\details{
The \code{minlossplot} function gives the best (lowest) C-V loss for each value of alpha.
}
