% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minlossplot.R
\name{minlossplot.penAFT.cva}
\alias{minlossplot.penAFT.cva}
\title{minlossplot}
\usage{
\method{minlossplot}{penAFT.cva}(x, ..., cv.type = c("min", "1se"), plot.it = c("min", "all"))
}
\arguments{
\item{x}{Object}

\item{cv.type}{For \code{minlossplot}, which cross-validated loss value to plot for each value of alpha.
This can be either \code{"min"} which is the minimum loss, or \code{"1se"} which is the highest loss within 1 standard error of the minimum.}
}
\description{
minloss desc
}
\details{
The \code{minlossplot} function gives the best (lowest) C-V loss for each value of alpha.
}
