% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penAFTUtils.r
\docType{package}
\name{penAFTutils}
\alias{penAFTutils}
\alias{penAFTutils-package}
\title{Utilities for penAFT}
\description{
Some quality-of-life functions to streamline the process of fitting elastic net models with the \code{penAFT} package, specifically:
\itemize{
\item \code{penAFT.cva} ??? .
\item \code{glmnet.formula} provides a formula/data frame interface to \code{glmnet}.
\item \code{cv.glmnet.formula} does a similar thing for \code{cv.glmnet}.
\item Methods for \code{predict} and \code{coef} for both the above.
\item A function \code{penAFT.cva} to choose both the alpha and lambda parameters via cross-validation, following the approach described in the help page for \code{penAFT.cv}. Optionally does the cross-validation in parallel.
\item Methods for \code{plot}, \code{predict} and \code{coef} for the above.
}
}
