% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.penAFT.R
\name{predict.penAFT.cva}
\alias{predict.penAFT.cva}
\title{\code{predict} method for objects of \code{penAFT.cva} class}
\usage{
\method{predict}{penAFT.cva}(
  object,
  newx,
  alpha,
  which = match(TRUE, abs(object$alpha - alpha) < 1e-08),
  lambda = NULL,
  ...
)
}
\arguments{
\item{object}{An object returned by \code{penAFT.cva}.}

\item{newx}{A matrix of predictors.}

\item{alpha}{desired value of alpha parameter (scalar)}

\item{which}{An alternative way of specifying alpha; the index number of the desired value.
within the \code{alpha} vector. If both \code{which} and \code{alpha} are supplied, the former takes precedence.}

\item{lambda}{if \code{lambda} is set to \code{NULL} the optimal value which minimizes C-V linear predictor scores is selected}

\item{...}{Further arguments to be passed to \code{penAFT::penAFT.predict()} function.}
}
\description{
\code{predict} method for objects of \code{penAFT.cva} class
}
\details{
The \code{predict} method computes predictions for a specific alpha value for a given \code{penAFT.cva} object.
in the object's stored \code{alpha} vector, and calls \code{penAFT::penAFT.predict} on the corresponding \code{penAFT.cv} fit. All the arguments to that function are (or should be) supported.
}
\seealso{
\link[penAFT:penAFT.predict]{penAFT::penAFT.predict}, \link[glmnet:predict.cv.glmnet]{glmnet::coef.cv.glmnet}
}
