% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runScript.R
\name{runScript}
\alias{runScript}
\title{Executes scripts stored in \code{cvaUtils} package}
\usage{
runScript(
  script = NA,
  package = "cvaUtils",
  subdir = "scriptsR4.1.3",
  echo = TRUE
)
}
\arguments{
\item{script}{character string containing name of the script to be executed. By default is set to NA.}

\item{package}{character string containing package name. By default \code{cvaUtils}.}

\item{subdir}{subdirectory containing scripts. By default: \code{scriptsR4.1.3}}

\item{echo}{logical. Used by source function. By default set to \code{TRUE}.}
}
\value{
Script is executed and results are printed.
}
\description{
Function excutes R script located in a specified subfolder. Default call of the function without arguments, prints a list of available scripts.
}
\examples{
runScript()

}
