% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvaUtils-package.R
\docType{package}
\name{cvaUtils-package}
\alias{cvaUtils}
\alias{cvaUtils-package}
\title{cvaUtils: Utility Functions for cva Objects}
\description{
Utility Functions for cva Objects
}
\details{
For a complete list of functions use:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ library(help = "cvaUtils")
}\if{html}{\out{</div>}}

or

\if{html}{\out{<div class="sourceCode">}}\preformatted{help.start()
}\if{html}{\out{</div>}}

and locate \code{cvaUtils} package
}
\author{
\strong{Maintainer}: Andrzej Galecki \email{agalecki@umich.edu} (\href{https://orcid.org/0000-0003-1542-4001}{ORCID})

}
\keyword{internal}
