% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossgrid.R
\name{lossgrid}
\alias{lossgrid}
\alias{lossgrid.default}
\alias{lossgrid.penAFT.cva}
\title{C-V loss}
\usage{
lossgrid(x, ...)

\method{lossgrid}{default}(x, ...)

\method{lossgrid}{penAFT.cva}(x, ..., cv.type = c("min", "1se"))
}
\arguments{
\item{x}{Object of}

\item{cv.type}{This can be either \code{"min"} which is the minimum loss, or \code{"1se"} which is the highest loss within 1 standard error of the minimum.}
}
\description{
\code{lossgrid} returns cross-validated loss for each value of alpha and lambda parameters from Elastic Net fit

Desc2

Desc3
}
\details{
The \code{lossgrid} is a generic function applied to cva objects gives C-V loss for each value
of alpha and lambda hyperparameters.
}
