% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossgrid.R
\name{lossgrid}
\alias{lossgrid}
\alias{lossgrid.default}
\title{lossgrid generic}
\usage{
lossgrid(x, ...)

\method{lossgrid}{default}(x, ...)
}
\arguments{
\item{x}{cva object}
}
\description{
lossgrid generic
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default
}}

