% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.penAFT.R
\name{predict.penAFT.cva}
\alias{predict.penAFT.cva}
\title{\code{predict} method for an object of \code{penAFT.cva} class}
\usage{
\method{predict}{penAFT.cva}(
  object,
  newx,
  alpha,
  which = match(TRUE, abs(object$alpha - alpha) < 1e-08),
  lambda = NULL,
  ...
)
}
\arguments{
\item{object}{For the \code{predict} and \code{coef} methods, an object returned by \code{penAFT.cva}.
predict method for objects of penAFT.cva class}

\item{newx}{For the \code{predict} method, a matrix of predictor variables.}

\item{alpha}{desired value of alpha parameter (scalar)}

\item{which}{An alternative way of specifying alpha; the index number of the desired value within the alpha vector. If both \code{which} and \code{alpha} are supplied, the former takes precedence.}

\item{...}{Further arguments to be passed to lower-level functions. In the case of \code{penAFT.cva}, these arguments are passed to \code{penAFT.cv}; for \code{predict} and \code{coef}, they are passed to \code{predict.cv.glmnet}; and for \code{plot} and \code{minlossplot}, to \code{plot}.}
}
\description{
\code{predict} method for an object of \code{penAFT.cva} class
}
\details{
The \code{predict} method computes predictions for a specific alpha value given a \code{penAFT.cva} object. It looks up the supplied alpha (possibly supplied indirectly via the \code{which} argument) in the object's stored \code{alpha} vector, and calls \code{penAFT::penAFT.predict} on the corresponding \code{penAFT.cv} fit. All the arguments to that function are (or should be) supported.
}
\seealso{
\link[penAFT:penAFT.predict]{penAFT::penAFT.predict}, \link[glmnet:predict.cv.glmnet]{glmnet::coef.cv.glmnet}
}
